Ported changes from GZDoom revision 363: - next round of reorganization.

+++ src/gl/common/a_dynlight.cpp
@@ -35,21 +35,18 @@
 **
 */
 
-#include "gl/gl_include.h"
-#include "gl/common/glc_renderer.h"
 #include "templates.h"
 #include "m_random.h"
 #include "r_main.h"
 #include "p_local.h"
 #include "c_dispatch.h"
 #include "g_level.h"
-#include "gl/gl_lights.h"
-#include "gl/gl_data.h"
-#include "gl/gl_basic.h"
-#include "gl/gl_functions.h"
-#include "gl/gl_values.h"
 #include "thingdef/thingdef.h"
 #include "i_system.h"
+#include "gl/common/glc_convert.h"
+#include "gl/common/glc_renderer.h"
+#include "gl/common/glc_dynlight.h"
+#include "gl/common/glc_templates.h"
 // [BB] New #includes.
 #include "network.h"
 
@@ -97,7 +94,7 @@ void AVavoomLight::BeginPlay ()
 
 void AVavoomLightWhite::BeginPlay ()
 {
-	byte l_args[5];
+	BYTE l_args[5];
 	memcpy(l_args, args, 5);
 	memset(args, 0, 5);
 	m_intensity[0] = l_args[0]*4;
@@ -196,7 +193,7 @@ void ADynamicLight::Activate(AActor *activator)
 	//Super::Activate(activator);
 	flags2&=~MF2_DORMANT;	
 
-	m_currentIntensity = m_intensity[0];
+	m_currentIntensity = float(m_intensity[0]);
 	m_tickCount = 0;
 
 	if (lighttype == PulseLight)
@@ -204,10 +201,10 @@ void ADynamicLight::Activate(AActor *activator)
 		float pulseTime = ANGLE_TO_FLOAT(this->angle) / TICRATE;
 		
 		m_lastUpdate = level.maptime;
-		m_cycler.SetParams(m_intensity[1], m_intensity[0], pulseTime);
+		m_cycler.SetParams(float(m_intensity[1]), float(m_intensity[0]), pulseTime);
 		m_cycler.ShouldCycle(true);
 		m_cycler.SetCycleType(CYCLE_Sin);
-		m_currentIntensity = (byte)m_cycler.GetVal();
+		m_currentIntensity = (BYTE)m_cycler.GetVal();
 	}
 }
 
@@ -261,10 +258,10 @@ void ADynamicLight::Tick()
 
 	case FlickerLight:
 	{
-		byte rnd = randLight();
+		BYTE rnd = randLight();
 		float pct = ANGLE_TO_FLOAT(angle)/360.f;
 		
-		m_currentIntensity = m_intensity[rnd >= pct * 255];
+		m_currentIntensity = float(m_intensity[rnd >= pct * 255]);
 		break;
 	}
 
@@ -277,7 +274,7 @@ void ADynamicLight::Tick()
 		
 		if (m_tickCount > ANGLE_TO_FLOAT(angle))
 		{
-			m_currentIntensity = m_intensity[0] + (amt * flickerRange);
+			m_currentIntensity = float(m_intensity[0] + (amt * flickerRange));
 			m_tickCount = 0;
 		}
 		break;
@@ -287,7 +284,7 @@ void ADynamicLight::Tick()
 	// These need some more work elsewhere
 	case ColorFlickerLight:
 	{
-		byte rnd = randLight();
+		BYTE rnd = randLight();
 		float pct = ANGLE_TO_FLOAT(angle)/360.f;
 		
 		m_currentIntensity = m_intensity[rnd >= pct * 255];
@@ -325,7 +322,7 @@ void ADynamicLight::Tick()
 	}
 
 	case PointLight:
-		m_currentIntensity = m_intensity[0];
+		m_currentIntensity = float(m_intensity[0]);
 		break;
 	}
 
@@ -365,7 +362,7 @@ void ADynamicLight::UpdateLocation()
 
 		if (lighttype == FlickerLight || lighttype == RandomFlickerLight) 
 		{
-			intensity = m_intensity[1];
+			intensity = float(m_intensity[1]);
 		}
 		else
 		{
